+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+### 5.2.31
+* fixed Hash Computations for non-primitive types
+* added `AsByteSpan` extension method for System.Array/string in >=net6.0
+
+### 5.2.30
+* removed UnsafeCoerce usages and several other net6.0+ fixes
+ 
 ### 5.2.29
 * Fixed color parsing to be independent of the current culture (regression in 5.2.27)
 * Added more value variants for Dictionary, Dict, and SymbolDict functions